Refactor findings read from FS to a separate include
